--- ios/utils/EventUtils.swift
@@ -1,14 +1,0 @@
-//
-//  TextUtils.swift
-//  CaptureProtection
-//
-//  Created by lethe on 3/3/25.
-//  Copyright © 2025 Facebook. All rights reserved.
-//
-
-import UIKit
-import Foundation
-
-public class EventUtils {
-    public static func eventMessage() {}
-}
